+++ AvgScript/src/webview/_onClickLink.ts
@@ -20,6 +20,15 @@ export async function handleOnClickLink(message: any) {
         const uri = vscode.Uri.parse(message.link);
         const line = parseInt(uri.fragment);
 
+        if (Number.isNaN(line)) {
+            const editor = await vscode.window.showTextDocument(uri
+                , {
+                    viewColumn: vscode.ViewColumn.Beside
+                });
+            editor.revealRange(new vscode.Range(0, 0, 0, 0)
+                , vscode.TextEditorRevealType.InCenterIfOutsideViewport);
+        }
+
         const doc = await vscode.workspace.openTextDocument(uri);
         const text = doc.lineAt(line).text;