+++ AvgScript/src/webview/dubList.ts
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/naming-convention */
 
 import { DubInfo } from "../functions/command";
-import { markDown_getLink, markDown_getMarkDownLevel, markDown_newLine, markdownParser } from "./_mdToHtml";
+import { markDown_getLink, markDown_getMarkDownLevel, markDown_linkEscape, markDown_newLine, markdownParser } from "./_mdToHtml";
 import { onClickLinkScript } from "./_onClickLink";
 
 export const narrator = '旁白';
@@ -22,9 +22,21 @@ export function dubList_getWebviewContent(dubMap: Map<string, DubInfo[]>) {
         + onClickLinkScript
         + `</html>`;
 
-    let markdown = markDown_getMarkDownLevel(1)
+    let markdown = '';
+
+    let title = markDown_getMarkDownLevel(1)
         + 'Dub List'
         + markDown_newLine;
+
+    const charListTitle = 'Character List';
+    let charList = markDown_getMarkDownLevel(2)
+        + charListTitle
+        + markDown_newLine;
+
+    const charListAppend = ',\t\t\t\t\t\t\t\t';
+    const charListWidth = 8;
+    let charListCount = 0;
+    const charListTitleLink = markDown_linkEscape(charListTitle);
 
     dubMap.forEach((value: DubInfo[], key: string) => {
         const templateInfo = value[0];
@@ -37,8 +49,12 @@ export function dubList_getWebviewContent(dubMap: Map<string, DubInfo[]>) {
             wordCount += dia.length;
 
             // diaContent += dia;
-            diaContent += markDown_getLink(dia
-                , info.uri + '#' + info.line.toString());
+            // diaContent += markDown_getLink(dia
+            //     , info.uri + '#' + info.line.toString());
+            diaContent += markDown_getLink('↪'
+                , info.uri + '#' + info.line.toString())
+                + ' '
+                + dia;
             diaContent += markDown_newLine;
         }
 
@@ -49,20 +65,43 @@ export function dubList_getWebviewContent(dubMap: Map<string, DubInfo[]>) {
             ? templateInfo.dialogueStruct.m_namePartRaw
             : narrator;
 
-        name += ': '
+        let charTitle = name + ': '
             + (internalName !== name
-                ? 'internal name: ' + internalName
+                ? 'internal name: ' + '\`' + internalName + '\`' + ', '
                 : '')
-            + 'line count: ' + value.length.toString()
+            + value.length.toString() + ' lines'
             + ', '
-            + 'word count: ' + diaContent.length.toString();
+            + wordCount.toString() + ' words';
+
+        const charTitleLink = markDown_linkEscape('↩ ' + charTitle);
+
+        // charList += markDown_getLink(name, '#' + charTitleLink.toLowerCase())
+        //     + markDown_newLine;
+
+        charList += markDown_getLink(name, '#' + charTitleLink.toLowerCase());
+        charListCount++;
+
+        if (charListCount >= charListWidth) {
+            charList += markDown_newLine;
+            charListCount = 0;
+        } else {
+            charList += charListAppend;
+        }
+
+        charTitle = markDown_getLink('↩', '#' + charListTitleLink.toLowerCase())
+            + ' '
+            + charTitle;
 
         markdown += markDown_getMarkDownLevel(2)
-            + name
+            + charTitle
             + markDown_newLine;
 
         markdown += diaContent;
     });
+
+    markdown = title
+        + charList.substring(0, charList.length - charListAppend.length) + markDown_newLine
+        + markdown;
 
     const html = markdownParser(markdown);